Close the security holes codex found, and the recommender bug behind them
Stored XSS, critical: uploads took Content-Type from the client and
/api/media served it back inline from this origin, so an uploader could
store markup under a key they legitimately own and run it with the
viewer's session attached. The type is now derived from the key against a
closed allowlist, with nosniff on every response. Three tests asserted the
old behaviour; one was literally named "uses the content type the client
stated".

IDOR: comments, reactions, playlist-add and the Shorts comment action
authenticated the caller and then trusted the id in the path. Any signed-in
account could act on a private video by guessing it. One shared
authorizeVideoAccess now answers, collapsing "not yours" into the same 404
as "no such video" — the discipline /api/media already followed for bytes.
A comment reaction also has to name the video it hangs off.

Private media no longer gets `public, max-age=31536000, immutable`: a
cached response is the one path where a correct authorisation check is
never consulted.

The recommender divided by the candidate's session count alone. That is
sound for ranking one seed and wrong the moment aggregateAcrossSeeds sums
across seeds — popular seeds dominated, and the stored score contradicted
the schema's own symmetry claim. Both counts now, in all three places that
state the rule. The incremental and full refresh paths had drifted apart on
exactly this; the formula is shared now rather than promised to match.

Expansion also credited a two-hop candidate to whichever seed reached it
first, halving the evidence for the candidate two of your videos lead to.

Every internal <a> is a next/link — 22 sites that were full page reloads.

Mutation-tested: the byte-writer, multi-seed attribution.
2,106 tests green.

+++ youtube/src/media/demux/sample-table.ts
@@ -385,11 +385,20 @@ export interface SampleTable {
   /** The `stts` delta, in the track's own timescale. */
   readonly durations: Uint32Array;
   /**
-   * Signed composition offsets, in the track's own timescale, or `undefined`
-   * when the track has no `ctts` — which is the ordinary case and means decode
-   * order is presentation order.
+   * Composition offsets, in the track's own timescale, or `undefined` when the
+   * track has no `ctts` — which is the ordinary case and means decode order is
+   * presentation order.
+   *
+   * `Float64Array` rather than `Int32Array`, and the reason is the version
+   * split {@link decodeCtts} exists to honour. A version-1 offset is `i32` and
+   * fits. A version-0 offset is **u32**, and its top half does not: an offset
+   * of `0x80000000` lands in an `Int32Array` as −2,147,483,648, turning a frame
+   * that should be presented far in the future into one presented far in the
+   * past. Storing both in a float keeps every legal value of either version
+   * exactly — u32 and i32 are both well inside 2^53 — instead of picking one
+   * version's width and silently corrupting the other's.
    */
-  readonly compositionOffsets: Int32Array | undefined;
+  readonly compositionOffsets: Float64Array | undefined;
   /**
    * 1 per sync sample, or `undefined` when there was no `stss` — which means
    * **every** sample is one. Use {@link isSyncSample} rather than reading this,
@@ -532,9 +541,9 @@ export function buildSampleTable(stbl: ParsedBox): SampleTable {
   /* ---- composition offsets ---- */
 
   const cttsBox = child(stbl, "ctts");
-  let compositionOffsets: Int32Array | undefined;
+  let compositionOffsets: Float64Array | undefined;
   if (cttsBox !== undefined) {
-    compositionOffsets = new Int32Array(sampleCount);
+    compositionOffsets = new Float64Array(sampleCount);
     let at = 0;
     for (const entry of decodeCtts(cttsBox).entries) {
       for (let i = 0; i < entry.sampleCount && at < sampleCount; i++) {
